성능개선 및 오류수정(서버이동시 클라이언트 leave이벤트 특정상황에서 서버주소 잘못나오던 부분)

+++ Include/Master.h
@@ -10,7 +10,7 @@ namespace Zero
 		Zero::String		m_Description;
 		int					m_ServerType;
 		Zero::NetAddress	m_Addr;
-		size_t				m_Clients;
+		int					m_Clients;
 
 		CMasterInfo()
 		{
